Add integration spec for non-forked tests
